Switch to IDs as relational references in Schema
This change is necessary to make things easier to refer to.
For our permission system, we are only handling the IDs and don't have
any relational information for those, like a name or similar.

Also, to make things reproducible, we introduced an optional id
attribute for each `add*` mutation, to be able to refer to a specific
relation (before, we were heavily reliant on the auto-increment
feature).

Now we have full control in which sequential order the datasets are
inserted and can be sure the dataset has the right ID.
This change also adapts the api-data.gql file.