fix: avoid a cron job running concurrently with itself
If the time taken for the cronjob to run is longer than the period
between jobs you can end up with dozens of cronjob processes running
inside the pod.

Since we already wrap the cron command, just add some locking to it so
that only one instance of a cron job can run at once.